feat(chart): bump version to 4.5.0 and add audit volume mount support (#118)
This pull request updates the `backend-service` Helm chart to version
4.5.0, introduces new functionality for audit volume configuration, and
updates the documentation accordingly.

### Version Update:
*
[`charts/backend-service/Chart.yaml`](diffhunk://#diff-6d4eea04b3a722161a691220447eb14e4d5d0e6a5098a90560a1cb1114beed8fL10-R10):
Updated the chart version from `4.4.0` to `4.5.0`.
*
[`charts/backend-service/README.md`](diffhunk://#diff-665570d6bdacaf97b119492af1bbbbdc25c5b4a952bd9d8f1341554d0dbf7661L7-R18):
Updated all references to the chart version from `4.4.0` to `4.5.0` in
the documentation and installation examples.

### New Feature - Audit Volume:
*
[`charts/backend-service/templates/extra-cronjobs.yaml`](diffhunk://#diff-17dddb96e3a53edaa5f41591c8f4da488cf23e3585269cd4fc380bb3bd399889R63-R66):
Added support for an optional `auditVolume` configuration, allowing
users to specify an audit volume mount path.